#50 Moved more transformation functions from coord.h to transforms.h
